requirements.txt
# Core dependencies
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
httpx>=0.25.0

# Scheduling and async
apscheduler>=3.10.0
asyncio>=3.4.3

# Web scraping and parsing
beautifulsoup4>=4.12.0
scrapy>=2.11.0
feedparser>=6.0.10

# Document processing
PyPDF2>=3.0.0
pypdf>=3.17.0
python-docx>=1.1.0
pillow>=10.1.0

# OCR and vision
pytesseract>=0.3.10
opencv-python>=4.8.0

# Email
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Calendar
icalendar>=5.0.11

# Messaging platforms
python-telegram-bot>=20.7
twilio>=8.11.0
slack-sdk>=3.26.0

# Knowledge management
notion-client>=2.2.1

# GitHub
PyGithub>=2.1.1
gitpython>=3.1.40

# Data processing
pandas>=2.1.4
numpy>=1.26.0

# Database
sqlalchemy>=2.0.23
chromadb>=0.4.18

# NLP and embeddings
sentence-transformers>=2.2.2
tiktoken>=0.5.2

# Network graph
networkx>=3.2.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# Code quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.1
